+++ feincms3/plugins/__init__.py
@@ -1,4 +1,5 @@
 from .external import *  # noqa
+from .html import *  # noqa
 from .richtext import *  # noqa
 from .snippet import *  # noqa
 from .versatileimage import *  # noqa