+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Change log
   exception instead of bubbling the last viewname's exception.
 - Added a ``RedirectMixin`` to ``feincms3.mixins`` for redirecting
   pages to other pages or arbitrary URLs.
+- Added a footgun plugin (raw HTML code).
 
 
 `0.11`_ (2016-09-19)